Remove wrapper traits from BaseResource + Rector rules for renames
BaseResource trimmed to the new shape:
- The 5 wrapper traits (ExtendsForms, ExtendsTables, ExtendsPages,
  ExtendsRelationManagers, ExtendsSubnavigation) are deleted.
- ExtendsForms / ExtendsTables had no remaining callers — every
  migrated resource overrides form() / table() directly and routes
  hooks through the {Resource}Form / {Resource}Table split classes.
- The structural hook methods (getPages / getRelations /
  getRecordSubNavigation) and their getDefault* slots are folded
  directly into BaseResource. Same hook names preserved
  (extendPages / getRelations / extendSubNavigation) since they
  have no Form/Table seam — the hook target stays the Resource.

BaseResource now matches Filament's own composed-trait pattern: thin
base that composes named concerns (HasLunarPermissions,
HasScoutGlobalSearch, ResolvesModelContract, CallsHooks) and owns the
three structural hook methods inline.

Three resources (StaffResource, TaxClassResource, TaxRateResource)
had their slim rewrite re-applied — earlier Write calls had errored
silently, leaving the legacy form unmigrated. Caught when the wrapper
traits stopped providing the fallback form() method.

Upgrade-package Rector additions:
- LunarSetList::V1_TO_V2_METHOD_RENAMES catalogues method renames on
  Lunar base classes.
- config/rector.php wires RenameMethodRector with MethodCallRename
  configurators from the new list.
- Initial entries: BaseResource::getDefaultForm → form and
  BaseResource::getDefaultTable → table, covering downstream resource
  subclasses that overrode the legacy slots.

Spec 0005 marked completed; TODO updated.
