+++ packages/admin/src/Filament/Resources/TaxClassResource.php
@@ -2,22 +2,15 @@
 
 namespace Lunar\Admin\Filament\Resources;
 
-use Awcodes\BadgeableColumn\Components\Badge;
-use Awcodes\BadgeableColumn\Components\BadgeableColumn;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Filament\Actions\EditAction;
-use Filament\Forms\Components\TextInput;
-use Filament\Forms\Components\Toggle;
-use Filament\Schemas\Components\Component;
-use Filament\Schemas\Components\Section;
+use Filament\Schemas\Schema;
 use Filament\Support\Facades\FilamentIcon;
 use Filament\Tables\Table;
-use Illuminate\Database\Eloquent\Model;
 use Lunar\Admin\Filament\Clusters\Taxes;
 use Lunar\Admin\Filament\Resources\TaxClassResource\Pages\CreateTaxClass;
 use Lunar\Admin\Filament\Resources\TaxClassResource\Pages\EditTaxClass;
 use Lunar\Admin\Filament\Resources\TaxClassResource\Pages\ListTaxClasses;
+use Lunar\Admin\Filament\Resources\TaxClassResource\Schemas\TaxClassForm;
+use Lunar\Admin\Filament\Resources\TaxClassResource\Tables\TaxClassTable;
 use Lunar\Admin\Support\Resources\BaseResource;
 use Lunar\Core\Models\Contracts\TaxClass as TaxClassContract;
 
@@ -46,71 +39,17 @@ class TaxClassResource extends BaseResource
         return FilamentIcon::resolve('lunar::tax');
     }
 
-    protected static function getMainFormComponents(): array
+    public static function form(Schema $schema): Schema
     {
-        return [
-            Section::make()->schema([
-                static::getNameFormComponent(),
-                static::getDefaultFormComponent(),
-            ]),
-        ];
+        return TaxClassForm::configure($schema);
     }
 
-    protected static function getNameFormComponent(): Component
+    public static function table(Table $table): Table
     {
-        return TextInput::make('name')
-            ->label(__('lunarpanel::taxclass.form.name.label'))
-            ->required()
-            ->maxLength(255)
-            ->autofocus();
+        return TaxClassTable::configure($table);
     }
 
-    protected static function getDefaultFormComponent(): Component
-    {
-        return Toggle::make('default')
-            ->label(__('lunarpanel::taxzone.form.default.label'));
-    }
-
-    public static function getDefaultTable(Table $table): Table
-    {
-        return $table
-            ->columns(static::getTableColumns())
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    protected static function getTableColumns(): array
-    {
-        return [
-            BadgeableColumn::make('name')
-                ->separator('')
-                ->suffixBadges([
-                    Badge::make('default')
-                        ->label(__('lunarpanel::taxclass.table.default.label'))
-                        ->color('gray')
-                        ->visible(fn (Model $record) => $record->default),
-                ])
-                ->label(__('lunarpanel::taxclass.table.name.label')),
-        ];
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getDefaultPages(): array
+    protected static function getDefaultPages(): array
     {
         return [
             'index' => ListTaxClasses::route('/'),
