+++ packages/admin/src/Support/Resources/BaseResource.php
@@ -2,13 +2,12 @@
 
 namespace Lunar\Admin\Support\Resources;
 
+use Filament\Pages\Page;
+use Filament\Resources\RelationManagers\RelationGroup;
+use Filament\Resources\RelationManagers\RelationManager;
+use Filament\Resources\RelationManagers\RelationManagerConfiguration;
 use Filament\Resources\Resource;
 use Lunar\Admin\Support\Concerns\CallsHooks;
-use Lunar\Admin\Support\Resources\Concerns\ExtendsForms;
-use Lunar\Admin\Support\Resources\Concerns\ExtendsPages;
-use Lunar\Admin\Support\Resources\Concerns\ExtendsRelationManagers;
-use Lunar\Admin\Support\Resources\Concerns\ExtendsSubnavigation;
-use Lunar\Admin\Support\Resources\Concerns\ExtendsTables;
 use Lunar\Admin\Support\Resources\Concerns\HasLunarPermissions;
 use Lunar\Admin\Support\Resources\Concerns\HasScoutGlobalSearch;
 use Lunar\Admin\Support\Resources\Concerns\ResolvesModelContract;
@@ -16,12 +15,42 @@ use Lunar\Admin\Support\Resources\Concerns\ResolvesModelContract;
 class BaseResource extends Resource
 {
     use CallsHooks;
-    use ExtendsForms;
-    use ExtendsPages;
-    use ExtendsRelationManagers;
-    use ExtendsSubnavigation;
-    use ExtendsTables;
     use HasLunarPermissions;
     use HasScoutGlobalSearch;
     use ResolvesModelContract;
+
+    public static function getPages(): array
+    {
+        return self::callStaticLunarHook('extendPages', static::getDefaultPages());
+    }
+
+    protected static function getDefaultPages(): array
+    {
+        return [];
+    }
+
+    /**
+     * @return array<class-string<RelationManager> | RelationGroup | RelationManagerConfiguration>
+     */
+    public static function getRelations(): array
+    {
+        return self::callStaticLunarHook('getRelations', static::getDefaultRelations());
+    }
+
+    protected static function getDefaultRelations(): array
+    {
+        return [];
+    }
+
+    public static function getRecordSubNavigation(Page $page): array
+    {
+        $pages = self::callStaticLunarHook('extendSubNavigation', static::getDefaultSubNavigation());
+
+        return $page->generateNavigationItems($pages);
+    }
+
+    protected static function getDefaultSubNavigation(): array
+    {
+        return [];
+    }
 }
