--- packages/admin/src/Support/Resources/Concerns/ExtendsTables.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Resources\Concerns;
-
-use Filament\Tables\Table;
-
-trait ExtendsTables
-{
-    public static function table(Table $table): Table
-    {
-        return self::callStaticLunarHook('extendTable', static::getDefaultTable($table));
-    }
-
-    protected static function getDefaultTable(Table $table): Table
-    {
-        return $table;
-    }
-}
